Atualizacao automática

+++ SB_FRAMEWORK/SBWebPaginas/src/main/java/com/super_bits/modulosSB/webPaginas/JSFBeans/SB/bean/BeanExemplo.java
@@ -92,7 +92,7 @@ public class BeanExemplo
 
     @NotNull
     @InfoCampo(tipo = FabCampos.PERCENTUAL)
-    private Double porcentagem;
+    private int porcentagem;
 
     @NotNull
     @InfoCampo(tipo = FabCampos.MOEDA_REAL)
